requirements.txt
allure_python_commons==2.9.45
ddddocr==1.4.3
opencv_python==4.5.5.64
openpyxl==3.0.10
Pillow==9.1.1
pytesseract==0.3.9
pytest==7.1.2
PyYAML==6.0
requests==2.27.1
